pyproject.toml
[project]
name = "cde-trading"
version = "0.1.0"
description = "Low-latency trading system for Coinbase Advanced Trade API"
requires-python = ">=3.11"
dependencies = [
    "coinbase-advanced-py>=1.8.0",
    "pyzmq>=26.0.0",
    "msgpack>=1.1.0",
    "uvloop>=0.21.0; sys_platform != 'win32'",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "pyarrow>=15.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
data-service = "data_service.main:main"
execution-service = "execution_service.main:main"
recorder = "recorder_service.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
